backend/requirements.txt
django
django-debug-toolbar
djangorestframework
djangorestframework-simplejwt
django-cors-headers
drf-spectacular
django-filter
pillow
channels
daphne
pytest
black
isort
pre-commit
mysqlclient
python-dotenv
